Refactor caching services into standalone classes.
The caching services previously existed in LocationService.cs. To improve the code organization and maintainability and conform with the Single Responsibility Principle, the MemoryCacheService and DistributedCacheServices are separated into their own standalone file. The LocationService is now only dependent on the interface definitions, avoiding the need to change the service if the caching strategy changes. This refactoring will not impact any existing functionality.
